feat(silo): refactor
