+++ src/silo/query_engine/actions/most_recent_common_ancestor.h
@@ -8,13 +8,13 @@
 
 #include <nlohmann/json_fwd.hpp>
 
-#include "silo/query_engine/actions/action.h"
+#include "silo/query_engine/actions/tree_action.h"
 #include "silo/query_engine/copy_on_write_bitmap.h"
 #include "silo/storage/table.h"
 
 namespace silo::query_engine::actions {
 
-class MostRecentCommonAncestor : public Action {
+class MostRecentCommonAncestor : public TreeAction {
   private:
    std::string column_name;
    bool print_nodes_not_in_tree;
@@ -24,10 +24,11 @@ class MostRecentCommonAncestor : public Action {
   public:
    MostRecentCommonAncestor(std::string column_name, bool print_nodes_not_in_tree);
 
-   arrow::Result<QueryPlan> toQueryPlanImpl(
-      std::shared_ptr<const storage::Table> table,
-      std::vector<CopyOnWriteBitmap> partition_filters,
-      const config::QueryOptions& query_options
+   arrow::Status addResponseToBuilder(
+      std::vector<std::string>& all_node_ids,
+      std::unordered_map<std::string_view, exec_node::JsonValueTypeArrayBuilder>& output_builder,
+      const storage::column::StringColumnMetadata* metadata,
+      bool print_nodes_not_in_tree
    ) const override;
 
    std::vector<schema::ColumnIdentifier> getOutputSchema(const schema::TableSchema& table_schema
