+++ src/silo/query_engine/actions/tree_action.h
@@ -1,0 +1,49 @@
+#pragma once
+
+#include <memory>
+#include <string>
+#include <vector>
+
+#include <arrow/result.h>
+
+#include <nlohmann/json_fwd.hpp>
+
+#include <arrow/compute/exec.h>
+#include "silo/query_engine/actions/action.h"
+#include "silo/query_engine/copy_on_write_bitmap.h"
+#include "silo/storage/table.h"
+
+#include "silo/query_engine/exec_node/arrow_util.h"
+#include "silo/query_engine/exec_node/json_value_type_array_builder.h"
+
+namespace silo::query_engine::actions {
+
+class TreeAction : public Action {
+  private:
+   std::string column_name;
+   bool print_nodes_not_in_tree;
+
+  public:
+   TreeAction(std::string column_name, bool print_nodes_not_in_tree);
+
+   std::vector<std::string> getNodeValues(
+      std::shared_ptr<const storage::Table> table,
+      const std::string& column_name,
+      std::vector<CopyOnWriteBitmap>& bitmap_filter
+   ) const;
+
+   virtual arrow::Status addResponseToBuilder(
+      std::vector<std::string>& all_node_ids,
+      std::unordered_map<std::string_view, exec_node::JsonValueTypeArrayBuilder>& output_builder,
+      const storage::column::StringColumnMetadata* metadata,
+      bool print_nodes_not_in_tree
+   ) const = 0;
+
+   arrow::Result<QueryPlan> toQueryPlanImpl(
+      std::shared_ptr<const storage::Table> table,
+      std::vector<CopyOnWriteBitmap> partition_filters,
+      const config::QueryOptions& query_options
+   ) const override;
+};
+
+}  // namespace silo::query_engine::actions
